requirements.txt
mysql-connector-python~=9.3.0
Flask~=3.0.2